I moved third party files around and fixed more warnings. I wanted to move third party code that gets compiled to be compiled with the framework.  I am now getting a warning due to linking the static lib for open gl in both the framework static lib and the game exe.  That will be fixed in the future for now I am taking a break

+++ Source/Framework/Graphics/Shape.cpp
@@ -34,7 +34,7 @@ void Shape::Init(std::vector<glm::vec2>& points, const glm::vec3& col)
 void Shape::Init(std::vector<glm::vec2>& points, const glm::vec3& col, GLenum primitive)
 {
 	m_Color = col;
-	m_NumVerts = points.size();
+	m_NumVerts = static_cast<int>(points.size());
 	m_PrimitiveType = primitive;
 
 	glGenBuffers(1, &m_Vbo);
